# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "سیستم مدیریت نوبت‌دهی کلینیک دندانپزشکی",
    "description": "API کامل و حرفه‌ای برای مدیریت نوبت‌دهی دو کلینیک دندانپزشکی\n\n⚠️ نکته مهم CSRF Protection:\n1. ابتدا از /auth/csrf-token یک token دریافت کنید\n2. در header درخواست‌های POST/PATCH/DELETE، X-CSRF-Token را با مقدار token ارسال کنید",
    "version": "1.4.0",
    "contact": {
      "name": "Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000/api",
      "description": "Development server"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Alternative port"
    }
  ],
...
  "paths": {
...
    "/faqs": {
      "get": {
        "tags": ["FAQs"],
        "summary": "لیست سوالات متداول",
        "operationId": "getFaqs",
        "parameters": [
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/Limit" },
          {
            "name": "published",
            "in": "query",
            "schema": { "type": "string", "enum": ["true", "false"] }
          }
        ],
        "responses": {
          "200": { "description": "لیست سوالات متداول" }
        }
      },
      "post": {
        "tags": ["FAQs"],
        "summary": "ایجاد سوال متداول (مدیر/منشی)",
        "operationId": "createFaq",
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["question", "answer"],
                "properties": {
                  "question": { "type": "string", "example": "آیا درمان ریشه درد دارد؟" },
                  "answer": { "type": "string", "example": "درمان ریشه با بی‌حسی موضعی انجام می‌شود و درد ندارد" },
                  "order": { "type": "integer", "example": 1 },
                  "published": { "type": "boolean", "example": true }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "سوال ایجاد شد" }
        }
      }
    },
    "/faqs/{id}": {
      "get": {
        "tags": ["FAQs"],
        "summary": "دریافت سوال متداول",
        "operationId": "getFaq",
        "parameters": [{ "$ref": "#/components/parameters/Id" }],
        "responses": {
          "200": { "description": "سوال متداول" }
        }
      },
      "patch": {
        "tags": ["FAQs"],
        "summary": "به‌روزرسانی سوال متداول (مدیر/منشی)",
        "operationId": "updateFaq",
        "security": [{ "cookieAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/Id" }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "question": { "type": "string" },
                  "answer": { "type": "string" },
                  "order": { "type": "integer" },
                  "published": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "سوال به‌روزرسانی شد" }
        }
      },
      "delete": {
        "tags": ["FAQs"],
        "summary": "حذف سوال متداول (مدیر/منشی)",
        "operationId": "deleteFaq",
        "security": [{ "cookieAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/Id" }],
        "responses": {
          "200": { "description": "سوال حذف شد" }
        }
      }
    },
...
    "/gallery": {
      "get": {
        "tags": ["Gallery"],
        "summary": "لیست تصاویر گالری",
        "operationId": "getGalleryImages",
        "parameters": [
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/Limit" },
          {
            "name": "published",
            "in": "query",
            "schema": { "type": "string", "enum": ["true", "false"] }
          }
        ],
        "responses": {
          "200": { "description": "لیست تصاویر گالری" }
        }
      },
      "post": {
        "tags": ["Gallery"],
        "summary": "آپلود تصویر گالری (مدیر/منشی)",
        "operationId": "uploadImage",
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["galleryImage"],
                "properties": {
                  "title": { "type": "string", "example": "تصویر کلینیک" },
                  "description": { "type": "string", "example": "توضیحات تصویر" },
                  "order": { "type": "integer", "example": 1 },
                  "published": { "type": "string", "enum": ["true", "false"] },
                  "galleryImage": { "type": "string", "format": "binary" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "تصویر آپلود شد" }
        }
      }
    },
    "/gallery/{id}": {
      "get": {
        "tags": ["Gallery"],
        "summary": "دریافت تصویر گالری",
        "operationId": "getGalleryImage",
        "parameters": [{ "$ref": "#/components/parameters/Id" }],
        "responses": {
          "200": { "description": "تصویر گالری" }
        }
      },
      "patch": {
        "tags": ["Gallery"],
        "summary": "به‌روزرسانی تصویر گالری (مدیر/منشی)",
        "operationId": "updateImage",
        "security": [{ "cookieAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/Id" }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "description": { "type": "string" },
                  "order": { "type": "integer" },
                  "published": { "type": "string", "enum": ["true", "false"] },
                  "galleryImage": { "type": "string", "format": "binary" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "تصویر به‌روزرسانی شد" }
        }
      },
      "delete": {
        "tags": ["Gallery"],
        "summary": "حذف تصویر گالری (مدیر/منشی)",
        "operationId": "deleteImage",
        "security": [{ "cookieAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/Id" }],
        "responses": {
          "200": { "description": "تصویر حذف شد" }
        }
      }
    },
...
  },
  "components": {
...
    "parameters": {
      "Id": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        },
        "description": "شناسه یکتای رکورد"
      },
      "Page": {
        "name": "page",
        "in": "query",
        "schema": {
          "type": "integer",
          "default": 1,
          "minimum": 1
        },
        "description": "شماره صفحه"
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "schema": {
          "type": "integer",
          "default": 10,
          "minimum": 1,
          "maximum": 100
        },
        "description": "تعداد رکوردها در هر صفحه"
      }
    },
...
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "dental.sid",
        "description": "Session-based authentication using cookies"
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "CSRF token برای حفاظت در برابر حملات CSRF. ابتدا /auth/csrf-token را صدا کنید و token دریافتی را در این header ارسال کنید"
      }
    }
  }
}